RandomGraph[{n, m}] returns a Graph with n vertices and m random edges
Implements RandomGraph[{n, m}] by sampling m edges without replacement
from the K_n edge set, and RandomGraph[{n, m}, k] by repeating the
draw k times. Requesting more edges than K_n supplies leaves the call
unevaluated, matching wolframscript.

+++ src/functions/math_ast/hypergeometric.rs
@@ -651,7 +651,7 @@ fn try_pfq_regularized_with_b_poles(
     })
     .max()
     .unwrap_or(0)
-    .max(0) as i64;
+    .max(0);
 
   // Compute the first non-zero term at n = min_n directly.
   let mut term: f64 = 1.0;
